refactor: move LicenseTypeDefinition to fitting namespace
