+++ src/Core/Liz.Core/Settings/ExtractLicensesSettingsBase.cs
@@ -1,4 +1,4 @@
-﻿using Liz.Core.License.Sources.LicenseType;
+﻿using Liz.Core.License.Contracts.Models;
 using System.Diagnostics.CodeAnalysis;
 
 namespace Liz.Core.Settings;
