+++ src/Nuke/Liz.Nuke.Tests/ExtractLicensesSettingsExtensionsTests.cs
@@ -1,5 +1,5 @@
 ﻿using FluentAssertions;
-using Liz.Core.License.Sources.LicenseType;
+using Liz.Core.License.Contracts.Models;
 using Nuke.Common.IO;
 using System.Reflection;
 using Xunit;
