+++ src/Nuke/Liz.Nuke/ExtractLicensesSettingsExtensions.cs
@@ -1,4 +1,4 @@
-﻿using Liz.Core.License.Sources.LicenseType;
+﻿using Liz.Core.License.Contracts.Models;
 using Nuke.Common.IO;
 
 namespace Liz.Nuke;
